Attach the [output] arg. of [create_for_stream] to the surface.
This allows, first, to make sure the value lives as long as the
surface (it is registered with the Garbage Collector for as long as
the surface is reachable).  Moreover, this behaves well with the
surface being passed around (for example with [get_target] which
returns the underlying surface of a context or [Surface.create_similar]).