Remove all "external" from the interface
Using "val" or "external" makes not difference to how ocamlopt will
compile the calls to the library functions.  Using "external" is
dangerous because it does not trigger the initialization code of the
library (which registers exceptions that, if used in the C code
without being registered, segfault the program).
See http://caml.inria.fr/mantis/view.php?id=6956